Removing Singleton.
Singletons are used for Dispersy and Database instances.  This is not
stricly necessary and, especially when stopping Dispersy, can be
difficult to deal with.

This patch removes all Singletons from the Dispersy core.

Since Dispersy.{has,get}_instance() can no longer be used to obtain
the Dispersy instance, this instance must be given as parameters when:

- a Community is created (create_community, join_community,
  get_master_members, load_community).

- a Member is created (should now use get_member,
  get_temporary_member_from_id, get_members_from_id, and
  get_member_from_database_id)

Similarly, since DispersyDatabase.{has,get}_instance() can no longer
be used, this should be replaced with dispersy.database.