Can again set the private key for an existing member.
The Member constructor is no longer called for each Member that is
returned or found in the cache (back when the cache was in
Member._cache).

Hence a new set_private_key(private_key) method is now available.
This method is called when the private_key is known but not yet
available in the cached Member instance.