Moved thread name parameter from start() to __init__().
This move is necessary to allow dispersy to start() the callback,
while not letting dispesy choose the name of the thread.

This patch also removes the prctl import that was used to prefix
"Tribler" to the thread name.  prctl is not part of standard python
hence I prefer not to use it, unless necessary.