nosetests behaviour keeps me confuddled.
Apparently the loggers (when no other loggers are enabled using -l)
are all enabled by default.  Hence, testing if it is enabled is not a
viable option.

Nosetests is designed for automated tests only, i.e. no parameters can
be passed to change the test behaviour.  As a work-around we are now
using an environmental variable to enable specific tests.