Removes MemberDestination.
Using MemberDestination can result in messages not being delivered
when the walker did not recently visit the member in question.  Using
CandidateDestination is preferred.

Currently only the signature request uses MemberDestination, we have
rewritten it to use CandidateDestination instead.  Since neither of
these destinations add anything to the binary packet, this change is
fully backwards compatible.