Renamed logger key 'hostname' into 'identifier'.
The key 'hostname' implies that it is the hostname of the machine, while
for a DAS4 run the hostname is not sufficient (you also need the peer
number).

By using more generic name 'identifier' we can use the hostname as a
default, while setting it to something else (i.e. the hostname + peer
number) when running on the DAS.